Rename ec_key to ec
